% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csodata.R
\name{cso_clear_cache}
\alias{cso_clear_cache}
\title{Clear csodata cache}
\usage{
cso_clear_cache()
}
\value{
Does not return a value, deletes the csodata cache.
}
\description{
Deletes all data cached by the csodata package. The cached data from the
csodata package is stored in a subdirectory of the default R.cache cache
at R.cache::getCachePath(). This function provides a quick way to delete
those files along with the directory to free up space.
}
\examples{
\dontrun{
cso_clear_cache()
}
}
