% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cso_get_meta.R
\name{cso_get_content}
\alias{cso_get_content}
\title{Returns a character vector listing the statistics in a CSO data table}
\usage{
cso_get_content(table_code, cache_data = TRUE)
}
\arguments{
\item{table_code}{string. A valid code for a table on data.cso.ie .}

\item{cache_data}{logical. Whether to use cached data, if available.
Default value is TRUE.}
}
\value{
character vector. The names of the statistics included in the
table, with one element for each statistic.
}
\description{
Returns a character vector listing the statistics in a CSO data table
}
\examples{
cso_get_content("EP008")
}
