% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cso_get_meta.R
\name{cso_get_var_values}
\alias{cso_get_var_values}
\title{Returns a list of the values of variables of a CSO data table}
\usage{
cso_get_var_values(table_code, cache_data = TRUE)
}
\arguments{
\item{table_code}{string. A valid code for one table from the
StatBank.}

\item{cache_data}{logical. Whether to use cached data, if available.
Default value is TRUE.}
}
\value{
list. It has length equal to the number of variables in the table,
and each element is a character vector which has all the values taken by
one variable.
}
\description{
Reads the table to determine all the unique values taken by the
variables in the table and returns them as a list.
}
\examples{
var_val <- cso_get_var_values("IPA03")
}
