% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cso_get_meta.R
\name{cso_disp_meta}
\alias{cso_disp_meta}
\title{Prints metadata from a PxStat table to the console}
\usage{
cso_disp_meta(table_code)
}
\arguments{
\item{table_code}{string. A valid code for a table on data.cso.ie .}
}
\value{
Does not return any values, rather the function prints the tables metadata to console.
}
\description{
Takes the output from \code{\link{cso_get_meta}} and prints it to the
console as formatted text.
}
\examples{
\dontrun{
cso_disp_meta("EP001")
}
}
