% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cso_get_meta.R
\name{cso_get_content}
\alias{cso_get_content}
\title{Returns a character vector listing the statistics in a CSO data table}
\usage{
cso_get_content(table_code)
}
\arguments{
\item{table_code}{string. A valid code for one table from the
StatBank.}
}
\value{
character vector. The names of the statistics included in the
table, with one element for each statistic.
}
\description{
Uses the ContentsIndicatorsList service of the CSO Statbank API to
returns a list the contained statistics for a table.
See \url{https://statbank.cso.ie/StatbankServices/StatbankServices.svc/
jsonservice/help/operations/ContentsIndicatorsList}
for more information on this.
}
\examples{
cso_get_content("EP008")
}
